[ADD] myinfo, payment 컴포넌트화 완료 - myinfo, payment 컴포넌트화 완료
