+++ src/screens/myPage/MyPage.js
@@ -1,10 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 import {Text, SafeAreaView, StyleSheet, ScrollView} from 'react-native';
-import AreaTit from '../../components/Texts/AreaTit';
-import StateButton from '../../components/Buttons/StateButton';
-import Inputs from '../../components/Inputs/Inputs';
-import AreaSubTit from '../../components/Texts/AreaSubTit';
+import SectionMyInfo from './SectionMyInfo';
+import SectionPayment from './SectionMyPayment';
 
 function MyPage() {
   return (
@@ -16,91 +14,9 @@ function MyPage() {
           </Header>
           <Contents>
             <Inner>
-              <SectionMyInfo>
-                <Wrapper>
-                  <TextArea>
-                    <AreaTit style={styles.title}>나의정보</AreaTit>
-                    <StageBox>
-                      <StateButton>수정</StateButton>
-                      <StateButton>수정</StateButton>
-                    </StageBox>
-                  </TextArea>
-                  <InputArea>
-                    <InputTit>아이디</InputTit>
-                    <InputBox>
-                      <Inputs placeholder="아이디" />
-                    </InputBox>
-                  </InputArea>
-                  <InputArea>
-                    <InputTit>이메일</InputTit>
-                    <InputBox>
-                      <Inputs placeholder="이메일" />
-                    </InputBox>
-                  </InputArea>
-                  <InputArea>
-                    <InputTit>이름</InputTit>
-                    <InputBox>
-                      <Inputs placeholder="이름" />
-                    </InputBox>
-                  </InputArea>
-                  <InputArea>
-                    <InputTit>휴대폰 번호</InputTit>
-                    <InputBox>
-                      <Inputs placeholder="휴대폰 번호" />
-                    </InputBox>
-                  </InputArea>
-                  <InputArea>
-                    <InputTit>주민번호</InputTit>
-                    <InputBox>
-                      <Inputs placeholder="앞 여섯자리 숫자" />
-                      <InputText>-</InputText>
-                      <Inputs placeholder="뒤 일곱자리 숫자" />
-                    </InputBox>
-                  </InputArea>
-                  <InputArea>
-                    <InputTit>주소</InputTit>
-                    <InputBox>
-                      <Inputs placeholder="주소" />
-                    </InputBox>
-                  </InputArea>
-                </Wrapper>
-              </SectionMyInfo>
+              <SectionMyInfo />
               {/* ---------------------------- */}
-              <SectionPayment>
-                <Wrapper>
-                  <TextArea>
-                    <AreaTit style={styles.title}>나의 결제 정보</AreaTit>
-                    <StageBox>
-                      <StateButton>수정</StateButton>
-                      <StateButton style={styles.removeBitton}>
-                        삭제
-                      </StateButton>
-                    </StageBox>
-                  </TextArea>
-                  <AreaTextBox>
-                    <AreaSubTit>카드별칭</AreaSubTit>
-                    <AreaSubTit>카드별칭</AreaSubTit>
-                  </AreaTextBox>
-                  <InputArea>
-                    <InputTit>이메일</InputTit>
-                    <InputBox>
-                      <Inputs placeholder="이메일" />
-                    </InputBox>
-                  </InputArea>
-                  <InputArea>
-                    <InputTit>이름</InputTit>
-                    <InputBox>
-                      <Inputs placeholder="이름" />
-                    </InputBox>
-                  </InputArea>
-                  <InputArea>
-                    <InputTit>휴대폰 번호</InputTit>
-                    <InputBox>
-                      <Inputs placeholder="휴대폰 번호" />
-                    </InputBox>
-                  </InputArea>
-                </Wrapper>
-              </SectionPayment>
+              <SectionPayment />
             </Inner>
           </Contents>
         </Container>
@@ -110,16 +26,6 @@ function MyPage() {
 }
 
 export default MyPage;
-
-const styles = StyleSheet.create({
-  title: {
-    color: '#69727b',
-  },
-  removeBitton: {
-    borderColor: '#ff6262',
-    color: '#ff6262',
-  },
-});
 
 const Container = styled.View`
   position: relative;
@@ -149,54 +55,3 @@ const Contents = styled.View`
 `;
 
 const Inner = styled.View``;
-
-const SectionMyInfo = styled.View`
-  background-color: #ffffff;
-  margin-bottom: 10px;
-`;
-
-const Wrapper = styled.View`
-  padding: 25px 20px 36px;
-`;
-
-const TextArea = styled.View`
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 9px;
-`;
-
-const StageBox = styled.View`
-  flex-direction: row;
-  align-items: center;
-`;
-
-const InputArea = styled.View`
-  padding: 7px 0;
-`;
-
-const InputTit = styled.Text`
-  padding-bottom: 6px;
-  font-size: 13px;
-  color: #48515a;
-  line-height: 20px;
-`;
-
-const InputBox = styled.View`
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-`;
-
-const InputText = styled.Text`
-  padding: 0 12px;
-  color: #48515a;
-`;
-
-const SectionPayment = styled(SectionMyInfo)``;
-
-const AreaTextBox = styled.View`
-  flex-direction: row;
-  align-items: center;
-  margin-top: 8px;
-`;
